Move form delete specs into separate spec file
The delete and destroy actions were moved out of the Forms controller in
a previous commit, we should move the specs as well.
